Adds re-export for controllers
