+++ routes/api/contacts.js
@@ -2,19 +2,20 @@ const express = require('express')
 
 const router = express.Router()
 
-const ctrl = require('../../controllers/contacts')
+const { contacts: ctrl } = require('../../controllers')
+const { ctrlWrapper } = require('../../helpers')
 const { isValidId } = require('../../middlewares')
 
-router.get('/', ctrl.listContacts)
+router.get('/', ctrlWrapper(ctrl.listContacts))
 
-router.get('/:contactId', isValidId, ctrl.getContactById)
+router.get('/:contactId', isValidId, ctrlWrapper(ctrl.getContactById))
 
-router.post('/', ctrl.addContact)
+router.post('/', ctrlWrapper(ctrl.addContact))
 
-router.put('/:contactId', isValidId, ctrl.updateContact)
+router.put('/:contactId', isValidId, ctrlWrapper(ctrl.updateContact))
 
-router.patch('/:contactId/favorite', isValidId, ctrl.updateStatusContact)
+router.patch('/:contactId/favorite', isValidId, ctrlWrapper(ctrl.updateStatusContact))
 
-router.delete('/:contactId', isValidId, ctrl.removeContact)
+router.delete('/:contactId', isValidId, ctrlWrapper(ctrl.removeContact))
 
 module.exports = router
